Keep guarded bkgsub output at the input precision
With bg_guard on, compute() returned float64 for a float32 mosaic, doubling
every SCI array written to disk. Measured on the real A2744 F444W mosaic
(56759 x 22171):

    _sci.fits   4.68 GiB -> 9.37 GiB   (BITPIX -32 -> -64)
    _i2d.fits  15.43 GiB -> 20.11 GiB

The promotion is entirely inside _cap_mesh, which reads the background mesh
as `np.asarray(bkg2d.background_mesh, dtype=float)` — Python float is
float64. Everything feeding it stays float32 (the s/sqrt(WHT) sigma map, the
multiscale ceiling, and Background2D.background itself), which is why the
unguarded path is unaffected and produced float32 all along. `sci - bkgmap`
then promotes.

Note that passing a different `dtype=` to the interpolator would NOT fix
this: in photutils' normal branch the zoom output follows the input array's
dtype, and `capped` is float64 before it ever gets there. That kwarg only
governs the degenerate ptp == 0 early return.

compute()'s own docstring promises "bkgd_subtracted : ndarray, sci -
background, same dtype as the input SCI", so this restores a documented
invariant rather than imposing a new policy. The arithmetic stays in float64
and only the result is cast; measured rounding cost is 2.4e-07 relative,
against drizzle inputs that are float32 to begin with.

tests/test_nircam_bkgsub_guard.py + tests/test_nircam_bkg.py: 23 passed.

+++ pipeline/campfire_pipeline/nircam/bkgsub.py
@@ -892,7 +892,16 @@ class SubtractBackground:
             log("Applying negativity guard (ceiling meshcap + trough pass)")
             off = (bitmask & 1) != 0
             bkgmap = self.apply_negativity_guard(sci, bkg, mask_final, off)
-        bkgd_subtracted = sci - bkgmap
+        # Cast back to the input precision. The guard works in float64 —
+        # _cap_mesh reads the mesh as `dtype=float` — so `sci - bkgmap` would
+        # otherwise promote a float32 mosaic to float64 and double every SCI
+        # array on disk (A2744 F444W: _sci 4.68 -> 9.37 GiB, _i2d 15.4 -> 20.1
+        # GiB), silently, for no gain over float32 drizzle inputs. The
+        # unguarded path already returns float32, and this method's docstring
+        # promises "same dtype as the input SCI"; without this the guard breaks
+        # that invariant. Keeping the arithmetic in float64 and casting only
+        # the result costs 2.4e-07 relative rounding.
+        bkgd_subtracted = (sci - bkgmap).astype(sci.dtype, copy=False)
 
         self.mask_final = mask_final
         return bkgd_subtracted, mask_final, bitmask
